+++ src/core/default.py
@@ -22,6 +22,8 @@ model_type = 'undefined'
 template_specifications = {}
 deformation_kernel_width = 1.0
 deformation_kernel_type = 'keops'
+deformation_kernel_device = 'auto'
+
 shoot_kernel_type = None
 number_of_time_points = 11
 concentration_of_time_points = 10