+++ src/core/models/geodesic_regression.py
@@ -35,6 +35,8 @@ class GeodesicRegression(AbstractStatisticalModel):
 
                  deformation_kernel_type=default.deformation_kernel_type,
                  deformation_kernel_width=default.deformation_kernel_width,
+                 deformation_kernel_device=default.deformation_kernel_device,
+
                  shoot_kernel_type=default.shoot_kernel_type,
                  concentration_of_time_points=default.concentration_of_time_points, t0=default.t0,
                  use_rk2_for_shoot=default.use_rk2_for_shoot, use_rk2_for_flow=default.use_rk2_for_flow,
@@ -71,7 +73,7 @@ class GeodesicRegression(AbstractStatisticalModel):
         # Deformation.
         self.geodesic = Geodesic(
             dense_mode=dense_mode,
-            kernel=kernel_factory.factory(deformation_kernel_type, deformation_kernel_width),
+            kernel=kernel_factory.factory(deformation_kernel_type, deformation_kernel_width, device=deformation_kernel_device),
             shoot_kernel_type=shoot_kernel_type,
             t0=t0, concentration_of_time_points=concentration_of_time_points,
             use_rk2_for_shoot=use_rk2_for_shoot, use_rk2_for_flow=use_rk2_for_flow)