+++ src/launch/compute_shooting.py
@@ -16,8 +16,11 @@ def compute_shooting(template_specifications,
                      dimension=default.dimension,
                      tensor_scalar_type=default.tensor_scalar_type,
                      tensor_integer_type=default.tensor_integer_type,
+
                      deformation_kernel_type=default.deformation_kernel_type,
                      deformation_kernel_width=default.deformation_kernel_width,
+                     deformation_kernel_device=default.deformation_kernel_device,
+
                      shoot_kernel_type=None,
                      initial_control_points=default.initial_control_points,
                      initial_momenta=default.initial_momenta,
@@ -34,7 +37,7 @@ def compute_shooting(template_specifications,
     Create the template object
     """
 
-    deformation_kernel = kernel_factory.factory(deformation_kernel_type, deformation_kernel_width)
+    deformation_kernel = kernel_factory.factory(deformation_kernel_type, deformation_kernel_width, device=deformation_kernel_device)
 
     (object_list, t_name, t_name_extension,
      t_noise_variance, multi_object_attachment) = create_template_metadata(