+++ tests/unit_tests/run_unit_tests.py
@@ -8,14 +8,14 @@ from unit_tests.test_api import API
 from unit_tests.test_array_readers_and_writers import ArrayReadersAndWritersTests
 from unit_tests.test_attachments import DistanceTests
 from unit_tests.test_auto_dimension import AutomaticDimensionDetectionTests
-from unit_tests.test_kernel_factory import KernelFactory, TorchKernel, KeopsVersusCuda
+from unit_tests.test_kernel_factory import KeopsVersusCuda, KernelFactoryTest, TorchKernelTest, KeopsKernelTest
 from unit_tests.test_parallel_transport import ParallelTransportTests
 from unit_tests.test_point_cloud import PointCloudTests
 from unit_tests.test_poly_line import PolyLineTests
 from unit_tests.test_shooting import ShootingTests
 from unit_tests.test_surface_mesh import SurfaceMeshTests
 
-TEST_MODULES = [API, KernelFactory, TorchKernel, KeopsVersusCuda,
+TEST_MODULES = [API, KernelFactoryTest, TorchKernelTest, KeopsKernelTest, KeopsVersusCuda,
                 ParallelTransportTests, DistanceTests, ArrayReadersAndWritersTests,
                 PolyLineTests, PointCloudTests, SurfaceMeshTests, ShootingTests,
                 AutomaticDimensionDetectionTests]