+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Bugfix: Gradients not computed when number_of_threads>1 and tensor_scalar_type is a FloatTensor. resolves #27
 - Split CLI commands into 2 groups: estimate and compute. resolves #6
 - Bugfix: memory leak when using estimate_longitudinal_atlas due to pytorch's autograd graph. resolves #33
+- Allow device selection when instantiation a kernel. 
+  'device' kwarg is now available when using the kernel factory and 'device-device' is now available when using the xml configuration file. resolves #13